+++ src/main/java/pl/com/bottega/ecommerce/sales/domain/invoicing/BookKeeper.java
@@ -16,10 +16,9 @@
 package pl.com.bottega.ecommerce.sales.domain.invoicing;
 
 import java.math.BigDecimal;
-import java.util.List;
 
-import pl.com.bottega.ecommerce.canonicalmodel.publishedlanguage.ClientData;
 import pl.com.bottega.ecommerce.canonicalmodel.publishedlanguage.Id;
+import pl.com.bottega.ecommerce.sales.domain.taxes.Tax;
 import pl.com.bottega.ecommerce.sharedkernel.Money;
 
 public class BookKeeper {
