+++ src/main/java/pl/com/bottega/ecommerce/sales/domain/invoicing/InvoiceLine.java
@@ -19,6 +19,7 @@
 package pl.com.bottega.ecommerce.sales.domain.invoicing;
 
 import pl.com.bottega.ecommerce.sales.domain.productscatalog.ProductData;
+import pl.com.bottega.ecommerce.sales.domain.taxes.Tax;
 import pl.com.bottega.ecommerce.sharedkernel.Money;
 
 public class InvoiceLine {
