+++ src/main/java/pl/com/bottega/ecommerce/sales/domain/taxes/Tax.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package pl.com.bottega.ecommerce.sales.domain.invoicing;
+package pl.com.bottega.ecommerce.sales.domain.taxes;
 
 import pl.com.bottega.ecommerce.sharedkernel.Money;
